Factorise rollup configurations
